+++ src/view/funciones/funciones-firebase.js
@@ -1,3 +1,4 @@
+/* eslint-disable prefer-promise-reject-errors */
 // ----------------------------- evento click de registro ------------------------------ //
 export const signup = (email, password) => firebase
   .auth()
@@ -47,45 +48,11 @@ export const getUserInfo = () => {
 };
 // ----------------------------- Inicio de sesión ------------------------------ //
 
-export const loginIn = (email, password) => {
-  firebase
-    .auth()
-    .signInWithEmailAndPassword(email, password)
-    .then((result) => {
-      const hash = '#/mainPage';
-      // Si el correo está verificado ingresa a la página(mainPage)
-      if (result.user.emailVerified) {
-        window.location.hash = hash;
-        console.log('verificado'); // eslint-disable-line
-        getUserInfo();
-      } else {
-        // De lo contrario su sesión se mantiene cerrado
-        firebase.auth().signOut();
-        const errorMessage = document.querySelector('#error-message');
-        errorMessage.style.display = 'inline';
-        errorMessage.textContent = `${result.user.displayName} por favor, realiza la verificación`;
-      }
-    })
-    .catch((error) => {
-      const errorMessage = document.querySelector('#error-message');
-      errorMessage.style.display = 'inline';
-      console.log(error); // eslint-disable-line
-      const errorCode = error.code;
-      if (errorCode === 'auth/invalid-email') {
-        errorMessage.textContent = 'Por favor ingrese su usuario y contraseña';
-        document.querySelector('#login-form').reset();
-      }
-      if (errorCode === 'auth/wrong-password') {
-        errorMessage.textContent = 'Contraseña incorrecta, inténtelo de nuevo';
-        document.querySelector('#login-form').reset();
-      }
-      if (errorCode === 'auth/user-not-found') {
-        errorMessage.textContent = 'El correo que ingresó no está registrado, por favor, regístrece';
-        document.querySelector('#login-form').reset();
-      }
-    });
-};
+export const loginIn = (email, password) => firebase
+  .auth()
+  .signInWithEmailAndPassword(email, password);
 
+// --------------------------- Inicio de sesión google --------------------------- //
 export const signInGoogle = () => {
   const providerGoogle = new firebase.auth.GoogleAuthProvider();
   const loginWithGoogle = firebase.auth().signInWithPopup(providerGoogle);
@@ -100,8 +67,8 @@ export const googlePromise = () => {
       window.location.hash = hash;
     })
     .catch((error) => {
-      console.log(error); // eslint-disable-line
-      console.log('no funciona'); // eslint-disable-line
+      console.log(error);
+      console.log('no funciona');
     });
 };
 // --------------------------- Inicio de sesión Facebook --------------------------- //
@@ -119,7 +86,7 @@ export const facebookPromise = () => {
       window.location.hash = hash;
     })
     .catch((error) => {
-  console.log(error); // eslint-disable-line
+      console.log(error);
     });
 };
 
@@ -129,13 +96,13 @@ export const signOut = () => firebase.auth().signOut();
 export const signOutPromise = () => {
   signOut()
     .then(() => {
-    console.log('cerraste sesión'); // eslint-disable-line
+      console.log('cerraste sesión');
       const hash = '#/login';
       window.location.hash = hash;
-      localStorage.clear(); // PRUEBA
+      localStorage.clear();
     })
     .catch((error) => {
-    console.log(error); // eslint-disable-line
+      console.log(error);
       const hash = '#/mainPage';
       window.location.hash = hash;
     });
@@ -146,12 +113,11 @@ export const signOutPromise = () => {
 export const publishPost = (objPublicacion) => new Promise((resolver, rechazar) => {
   firebase.firestore().collection('posts').add(objPublicacion)
     .then((docRef) => {
-      console.log('Documento escrito con el ID: ', docRef.id); // eslint-disable-line
+      console.log('Documento escrito con el ID: ', docRef.id);
       resolver('documeto registrado');
     })
     .catch((error) => {
-      console.log(error); // eslint-disable-line
-      // eslint-disable-next-line prefer-promise-reject-errors
+      console.log(error);
       rechazar('documeto no registrado');
     });
 });
@@ -182,8 +148,7 @@ export const likepublish = (idPost) => {
     }
   })
     .catch((error) => {
-  console.log(error); // eslint-disable-line
-
+      console.log(error);
     });
 };
 
@@ -191,12 +156,10 @@ export const likepublish = (idPost) => {
 export const deletePost = (idPost) => new Promise((resolver, rechazar) => {
   const post = firebase.firestore().collection('posts').doc(idPost);
   post.delete().then(() => {
-    console.log('Document successfully deleted!'); // eslint-disable-line
     resolver('eliminado');
   })
     .catch((error) => {
-      console.error('Error removing document: ', error); // eslint-disable-line
-      // eslint-disable-next-line prefer-promise-reject-errors
+      console.error('Error removing document: ', error);
       rechazar('no se pudo eliminar');
     });
 });
@@ -204,16 +167,13 @@ export const deletePost = (idPost) => new Promise((resolver, rechazar) => {
 // ------mostrar los like de los usuarios-------//
 export const showlike = (idPost) => new Promise((resolver, reject) => {
   const showPostLike = firebase.firestore().collection('posts').doc(idPost);
-  // showlike(idPost).get().then((res)
   showPostLike.get().then((res) => {
     if (res.exists) resolver(res);
     else {
-      // eslint-disable-next-line prefer-promise-reject-errors
       reject(null);
     }
   })
     .catch(() => {
-      // eslint-disable-next-line prefer-promise-reject-errors
       reject(null);
     });
 });
