+++ src/view/login-view.js
@@ -1,4 +1,6 @@
-import { loginIn, facebookPromise, googlePromise } from './funciones/funciones-firebase.js';
+import {
+  loginIn, getUserInfo, facebookPromise, googlePromise,
+} from './funciones/funciones-firebase.js';
 
 export const viewLogin = () => {
   const loginSection = `
@@ -37,7 +39,39 @@ document.addEventListener('click', (e) => {
   if (e.target.id === 'signin-btn') {
     const email = document.querySelector('#email').value;
     const password = document.querySelector('#password').value;
-    loginIn(email, password);
+    loginIn(email, password)
+      .then((usuario) => {
+        const hash = '#/mainPage';
+        // Si el correo está verificado ingresa a la página(mainPage)
+        if (usuario.user.emailVerified) {
+          window.location.hash = hash;
+          getUserInfo();
+        } else {
+        // De lo contrario su sesión se mantiene cerrado
+          firebase.auth().signOut();
+          const errorMessage = document.querySelector('#error-message');
+          errorMessage.style.display = 'inline';
+          errorMessage.textContent = `${usuario.user.displayName} por favor, realiza la verificación`;
+        }
+      })
+      .catch((error) => {
+        const errorMessage = document.querySelector('#error-message');
+        errorMessage.style.display = 'inline';
+      console.log(error); // eslint-disable-line
+        const errorCode = error.code;
+        if (errorCode === 'auth/invalid-email') {
+          errorMessage.textContent = 'Por favor ingrese su usuario y contraseña';
+          document.querySelector('#login-form').reset();
+        }
+        if (errorCode === 'auth/wrong-password') {
+          errorMessage.textContent = 'Contraseña incorrecta, inténtelo de nuevo';
+          document.querySelector('#login-form').reset();
+        }
+        if (errorCode === 'auth/user-not-found') {
+          errorMessage.textContent = 'El correo que ingresó no está registrado, por favor, regístrece';
+          document.querySelector('#login-form').reset();
+        }
+      });
   }
 });
